gev2: merged release 4.4 branch (rev. 55104)
git-svn-id: http://svn.ilias.de/svn/ilias/branches/CaT/releases/4_4_generali2@55115 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/TestQuestionPool/classes/class.assQuestionGUI.php
@@ -309,8 +309,7 @@ abstract class assQuestionGUI
 		$page_gui->setQuestionHTML(array($this->object->getId() => $html));
 		$page_gui->setOutputMode("presentation");
 		$presentation = $page_gui->presentation();
-		// bugfix for non XHTML conform img tags in ILIAS Learning Module Editor
-		$presentation = preg_replace("/src=\".\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
+		$presentation = preg_replace("/src=\"\\.\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
 		return $presentation;
 	}
 
@@ -340,7 +339,7 @@ abstract class assQuestionGUI
 		
 		if( $this->object->areObligationsToBeConsidered() && ilObjTest::isQuestionObligatory($this->object->getId()) )
 		{
-			$obligatoryString = ' *';
+			$obligatoryString = '([-_-])';
 		}
 		else
 		{
@@ -371,8 +370,13 @@ abstract class assQuestionGUI
 		}
 		$presentation = $page_gui->presentation();
 		if (strlen($maxpoints)) $presentation = str_replace($maxpoints, "<em>$maxpoints</em>", $presentation);
-		// bugfix for non XHTML conform img tags in ILIAS Learning Module Editor
-		$presentation = preg_replace("/src=\".\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
+		if (strlen($obligatoryString))
+		{
+			$replacement	='<br><span class="obligatory" style="font-size:small">'.
+				$this->lng->txt("tst_you_have_to_answer_this_question").'</span>';
+			$presentation 	= str_replace($obligatoryString, $replacement, $presentation);
+		}
+		$presentation = preg_replace("/src=\"\\.\\//ims", "src=\"" . ILIAS_HTTP_PATH . "/", $presentation);
 		return $presentation;
 	}
 	
@@ -1223,12 +1227,15 @@ abstract class assQuestionGUI
 				$file->setRequired(TRUE);
 				$file->enableFileNameSelection("filename");
 				//$file->setSuffixes(array("doc","xls","png","jpg","gif","pdf"));
-				if ($_FILES["file"]["tmp_name"])
+				if( $_FILES["file"]["tmp_name"] && $file->checkInput() )
 				{
 					if (!file_exists($this->object->getSuggestedSolutionPath())) ilUtil::makeDirParents($this->object->getSuggestedSolutionPath());
+					
 					$res = ilUtil::moveUploadedFile($_FILES["file"]["tmp_name"], $_FILES["file"]["name"], $this->object->getSuggestedSolutionPath() . $_FILES["file"]["name"]);
 					if ($res)
 					{
+						ilUtil::renameExecutables($this->object->getSuggestedSolutionPath());
+						
 						// remove an old file download
 						if (is_array($solution_array["value"])) @unlink($this->object->getSuggestedSolutionPath() . $solution_array["value"]["name"]);
 						$file->setValue($_FILES["file"]["name"]);
@@ -1246,6 +1253,7 @@ abstract class assQuestionGUI
 					}
 					else
 					{
+						// BH: $res as info string? wtf? it holds a bool or something else!!?
 						ilUtil::sendInfo($res);
 					}
 				}
